split services into generalServices
