+++ client/src/features/log-page/LogPage.tsx
@@ -2,7 +2,7 @@ import type React from "react";
 import { useState, useEffect } from "react";
 
 export default function LogPage () {
-    
+
     return(
         <div>
