+++ server/services/kubernetesService.js
@@ -1,8 +1,6 @@
 // Lines 2 - 33 are basic kubernetes API setup
 import * as k8s from '@kubernetes/client-node';
 import dotenv from 'dotenv';
-import fs from 'fs';
-import path from 'path';
 process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0';
 dotenv.config();
 // Defines helper functions that will connect middleware to the Kubernetes API Client functions
@@ -115,33 +113,6 @@ const kubernetesService = {
             }
         }
     },
-    //function that checks if the environment variables exist and if the env file exists, otherwise create it
-    checkEnv: () => {
-        // console.log(process.env.KUBERNETES_SERVER);
-        // console.log(process.env.KUBERNETES_TOKEN);
-        //console.log('envcheck run');
-        if (!process.env.KUBERNETES_SERVER || !process.env.KUBERNETES_TOKEN) {
-            const envPath = path.resolve(path.resolve('./.env'));
-            if (!fs.existsSync(envPath)) {
-                const defaultEnv = 'KUBERNETES_SERVER=\n' + 'KUBERNETES_TOKEN=';
-                fs.writeFileSync(envPath, defaultEnv.trim());
-                return 'init';
-            }
-            else {
-                return 'noVar';
-            }
-        }
-        else {
-            return 'exist';
-        }
-    },
-    writeEnv: (key, address) => {
-        const envPath = path.resolve(path.resolve('./.env'));
-        const fileEnv = 'KUBERNETES_SERVER=https://' + address + '\n' + 'KUBERNETES_TOKEN=' + key;
-        fs.writeFileSync(envPath, fileEnv, 'utf-8');
-        process.env.KUBERNETES_SERVER = 'https://' + address;
-        process.env.KUBERNETES_TOKEN = key;
-    }
 };
 // Exports service object for use as helper functions
 export default kubernetesService;
